TradingBeast8_Backend/requirements.txt
MetaTrader5==5.0.37
python-dotenv==1.0.1
requests==2.31.0